Initial upload of 3DMark06 test-profile
This test-profile assumes you own a 3DMark06 professional edition key.
If not so, then this test won't be useful as it relies on the command line options
the professional edition offers for automation.
During the installation of the test you need to have an environmental variable named key containing your 3DMark06 Professional Edition key, so ahk can read it and register your product automatically.
You don't need the variable set for future runs of the test-profile. Only during the installation process.